+++ src/settings/MappingProfiles/initialDetails/MARC_BIBLIOGRAPHIC.js
@@ -1,20 +1,7 @@
 const MARC_BIBLIOGRAPHIC = {
   name: 'marcBib',
   recordType: 'MARC_BIBLIOGRAPHIC',
-  mappingFields: [{
-    name: 'discoverySuppress',
-    enabled: true,
-    path: 'marcBib.discoverySuppress',
-    value: null,
-    booleanFieldAction: 'IGNORE',
-    subfields: [],
-  }, {
-    name: 'hrid',
-    enabled: true,
-    path: 'marcBib.hrid',
-    value: '',
-    subfields: [],
-  }],
+  marcMappingDetails: [{ order: 0 }],
 };
 
 export default MARC_BIBLIOGRAPHIC;