+++ test/bigtest/mocks/form-config-samples.js
@@ -4728,6 +4728,7 @@ export const formConfigSamples = [
           }, {
             controlType: 'MARCTable',
             staticControlType: 'Fragment',
+            id: 'marc-modification-table',
             dataAttributes: {},
           }],
         }],